requirements.txt
fastapi==0.58.0
python-gitlab==2.3.1
uvicorn==0.11.5